fix : separate login and register
