+++ src/controllers/auth/register.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from 'express';
 
-import { User } from '../models/User';
-import { sendRegistrationEmail } from '../utils/emailService';
-import { validateUser } from '../middlewares/validationMiddleware';
-import {generateToken} from '../utils/generateToken'
+import { User } from '../../models/User';
+import { sendRegistrationEmail } from '../../utils/emailService';
+import { validateUser } from '../../middlewares/validationMiddleware';
+import {generateToken} from '../../utils/generateToken';
 
 
 export const registerUser = [
@@ -39,27 +39,3 @@ export const registerUser = [
         }
     },
 ];
-
-export const loginUser = async (req: Request, res: Response) => {
-    const { email, password } = req.body;
-
-    try {
-        const user = await User.findOne({ email });
-
-        if (!user || !(await user.matchPassword(password))) {
-            return res.status(400).json({ error: { message: 'Invalid credentials', code: 400 } });
-        }
-        const token = generateToken(user.id);
-
-        res.json({
-            result: {
-                id: user.id,
-                email: user.email,
-                storeName: user.storeName,
-                token,
-            },
-        });
-    } catch (err) {
-        res.status(500).json({ error: { message: 'Error Occurred!', code: 500 } });
-    }
-};
